fix all component names
