+++ src/Header/HeaderBurger.scss
@@ -1,0 +1,73 @@
+.burger-menu {
+ 
+  
+  position: absolute;
+  width: 100%;
+  height: 260px;
+  transition: right 0.3s linear;
+  right: -150%;
+  
+  top: 80px;
+  z-index: 999;
+  background-color: rgb(0, 0, 0);
+  
+}
+
+.is-nav-open {
+  right:  0;
+  display: block;
+  
+}
+
+
+
+.nav-list_item--burger {
+  padding-top: 15px;
+  border-bottom: 1px solid #cccccc;
+  padding-left: 25px;
+  padding-bottom: 15px;
+}
+
+
+
+.menu {
+  display: none;
+  margin-right: 20px;
+  width: 24px;
+  height: 24px;
+  position: relative;
+
+  span {
+    height: 2px;
+    width: 100%;
+    display: block;
+    background-color: #ffffff;
+    position: absolute;
+    left: 0;
+    top: 50%;
+    margin-top: -1px;
+  }
+
+  &::before,
+  &::after {
+    content: "";
+    height: 2px;
+    width: 100%;
+    display: block;
+    background-color: #ffffff;
+    position: absolute;
+    left: 0;
+  }
+
+  &::before {
+    top: 0;
+  }
+
+  &::after {
+    bottom: 0;
+  }
+
+  @media (max-width: 768px) {
+    display: block;
+  }
+}
