+++ src/Services/Services.js
@@ -3,7 +3,7 @@ import MainContainer from '../Containers/Containers'
 import Centering from '../Containers/Centering-container'
 import Fade from 'react-reveal/Fade'
 import { Element } from 'react-scroll'
-import './services.scss'
+import './Services.scss'
 import iconLike from '../icons/like.svg'
 import iconWorld from '../icons/world.svg'
 import iconDesktop from '../icons/desktop.svg'
